requirements.txt
asgiref==3.5.0
attrs==21.4.0
blessings==1.7
bpython-django==0.22.11
certifi==2022.6.15
charset-normalizer==2.1.0
curtsies==0.3.10
cwcwidth==0.1.6
Django==3.2.10
django-extensions==3.2.0
djangorestframework==3.14.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
mysqlclient==2.1.0
packaging==21.3
param==1.12.0
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.7
pytest==7.1.2
pytest-django==4.5.2
pytz==2021.3
pyxdg==0.28
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
uWSGI==2.0.21
uwsgitop==0.11

